Add `-location` utility target to generate the location of a target's artifacts
Summary:
Add `-location` utility target to generate the location of a target's
artifacts.  The utility for this is to provide a mechanism for `python_*`
targets to discover the location of dependent artifacts in the `buck-out/`
path.  This mechanism is intended to be used instead of the current `__file__`
based method of discovering bundled resources to support OSS builds.  In OSS
the python packaging is done via `pex` and by default `pex` files are marked as
`zip-safe` which means that the native python import logic can import resources
directly from the `pex` zip file without unpacking.

Note that we *can* force `not-zip-safe` `pex` files, which means that they get
unpacked onto disk in all cases, but my preference would be that we not require
that behavior.  This mechanism provides an working alternative to forcing a
specific runtime behavior.

Note also that this works well for binaries/unittest that only run in the context of
a buck working directory, ie: unittests.  This won't be that useful for binaries
that are assumed to be bundled together such as `yum-from-snapshot`, which
embeds `repo-server` inside of it.

Differential Revision: D18783471

fbshipit-source-id: 0ce643d9130e35a40faeaa6ef5129cb3b9c60546