Remove trailing slash for rpm sources
Summary:
A recent version of buck that was just released does not like the
trailing `/` on the `sources` attribute of an `rpmbuild_layer`.  This `sources`
attribute ultimately makes it's way to `maybe_export_file` and then
`export_file` where it blows up with the trailing slash.  As such, since 12/4
at ~9:00 AM PDT, `fs_image` tests on trunk have been broken.

It appears that the
reason for the trailing slash, to ensure that the top-level directory of the
`sources` is not included in the tarball, is no longer an issue.  See the buck
output of the tarball being constructed by this test:

```
]# tar tvf buck-out/gen/fs_image/compiler/tests/toy-rpm-build-source.tgz/toy-rpm-build-source.tgz
drwxr-xr-x 0/0               0 2018-01-01 00:00 ./
-rw-r--r-- 0/0              40 2018-01-01 00:00 ./toy_src_file
```

Reviewed By: snarkmaster

Differential Revision: D18818936

fbshipit-source-id: 2f00e290da628d3a01b41aa830c36f0d7253caa3