Add utility to generate the location of a target's artifacts for use within unittests (#5)
Summary:
Pull Request resolved: https://github.com/facebookincubator/fs_image/pull/5

Add a macro that will generate the location of a target's
artifact.  The utility for this is to provide a mechanism for `python_*`
targets to discover the location of dependent artifacts in the `buck-out/`
path.  This mechanism is intended to be used instead of the current `__file__`
based method of discovering bundled resources to support OSS builds.  In OSS
the python packaging is done via `pex` and by default `pex` files are marked as
`zip-safe` which means that the native python import logic can import resources
directly from the `pex` zip file without unpacking.

Note that we *can* force `not-zip-safe` `pex` files, which means that they get
unpacked onto disk in all cases, but my preference would be that we not require
that behavior.  This mechanism provides an working alternative to forcing a
specific runtime behavior.

Note also that this works well for binaries/unittest that only run in the context of
a buck working directory, ie: unittests.  This won't be that useful for binaries
that are assumed to be bundled together such as `yum-from-snapshot`, which
embeds `repo-server` inside of it.

Reviewed By: snarkmaster

Differential Revision: D18783471

fbshipit-source-id: 66fc3829c590572c12e214f7f40ce921b59a52e0