`s/install_executable/install_buck_runnable/`
Summary:
This is a pure-rename diff, but it contains the rationale for the changes that follow.

The original design for `image.install*` items conflated "is the target `buck run`nable?" with "should executable bits be set on the copy inside the layer?".

These are separate concerns. For example, if one checks in an executable `.sh` into the repo, it is not a `buck run`nable target, and does not need our mode/dev target wrapping. However, the user **will** want executable bits to be set on the destination.

Today, they have to write `image.install_data("a.sh", "/a.sh", mode=0o555)`. After this stack, they will just need `image.install("a.sh", "/a.sh")`.

Reviewed By: filbranden

Differential Revision: D18904751

fbshipit-source-id: 3f40b19b68c11f5dd912bbc57f16b3364a2fa91d