Change `mode` semantics of `image.install`
Summary:
It was a poor decision to use "is buck runnable" to determine whether the installed file will be executable. As noted on D18904751, this breaks the common case of executable scripts.

Instead, we will test if a source file is executable by the repo user, and install it accordingly.

Reviewed By: filbranden

Differential Revision: D18905604

fbshipit-source-id: 972a2a5cec02aaaf0aae17e7538a8536d8f93c70