Replace wrapper script with a direct invocation of bash
Summary:
Use `/bin/bash -c '...'` to run our script inside the nspawn container.

This way, we don't need to drop a file inside the container filesystem while booting it for testing.

(Note that we still touch the filesystem by creating an `/outerproc` directory. We can later discuss possible approaches to prevent creation of that directory as well.)

Reviewed By: snarkmaster

Differential Revision: D18941015

fbshipit-source-id: 41ee4931d83c4bfaee7bcbca8a8de823eda816a9