Support writing `RepoSnapshot` as SQLite
Summary:
Instead of having per-repo files `{rpm,repodata,repomd}.json`, I want to have a single `snapshot.sql3`, with as many indexes as I want.

There are only a couple of differences in how the data are stored:
 - The SQLite contains all the new fields (NVREA & source_rpm) added by the previous diff.
 - For `mutable_rpm` errors, the `storage_id` is populated as a column, one does not need to reach into `error_json` to fetch it. As a consequence, `repo_server.py` will need to refuse to serve rows with populated errors.

Since this is a DB, indexes are easy to specify.

- The NVREA index is meant for "allowable version" work.

- The `repo, path` index is (roughly) what `repo_server.py` might need to avoid loading the DB into RAM (aside: it's lame that we have to look up a path in 3 tables, so that might be worth improving later). The stacked `repo_server.py` implementation actually does not use this index, so it could be deleted and NVREA made primary for a DB size win.

You will note that I'm not running `VACUUM;` here. I tried it on a big DB, and the change in size was 0.3%, so I left it out.

It's also important to note that schema changes are **cheap** for us — the code that uses a snapshot is source-controlled together with the snapshot itself, so it's not a big deal to land a change to both at once.

I **deliberately** left out `yum.conf` and `gpg_keys` out of the DB. These change infrequently, and owing to their importance, we really do want source-control history for both these to be easily inspectable.

Reviewed By: zeroxoneb

Differential Revision: D18878829

fbshipit-source-id: da6101075e76c413a8f274b15a4bdc2fdcfeafe6