Add third_party.library macros to oss_shim
Summary:
This adds a simple `third_party` interface for the `oss_shim.bzl` that will
allow the OSS implementation of this to support third-party library references
in a similar way as is done internally.  The goal is to minimize the varation
of this vs standard internal tools with the hope that future mistakes or
misuses can be minimised.

The new interface is:

- `third_party.library(spec)`:  Returns a target string for the
requested
third-party library.

Reviewed By: snarkmaster

Differential Revision: D18666643

fbshipit-source-id: d4584a885985425238de899578290ddc6cab846d