Target to prefetch `snapshot.sql3` for snapshot
Summary:
Many different parts of the build need to access the DB. It would be quite inefficient for them to independently fetch the sqlite DB from storage. Instead, let's have Buck cache it for us.

A stacked diff will make it so that the compiler takes the output of the `multi_repo_snapshot` target as an argument. This is a very useful change from today, where the snapshot is embedded in `yum-from-snapshot`.

Reviewed By: zeroxoneb

Differential Revision: D18878833

fbshipit-source-id: 4cc5ea407cc4510c0b49cef703f92ceff9a51640