Include dependencies for non-facebook plugins in rpm subsystem
Summary:
This includes additional dependencies for non-facebook plugins in the
rpm subsystem.  Yes, these non-fb deps are effectively redundant because the fb
deps end up bringing in the vanilla ones.   However, this is needed to support
OSS builds.  In OSS, any FB specific dep will be stripped from a target.  In
theory, we could instead *mutate* the fb specific dep to become non-fb specific,
however this does not work for `deleted_mutable_rpms` as the mutated dep would
become a self-reference. In order to support that specific case the OSS shim
would need to be aware of it, and that's not something we want.  The generic
approach of 'strip all FB specific deps' I think is the right way to go for now.

In the future we could introduce another layer in the bzl that provides some kind
of 'plugin implementation chooser' for `python_binary` targets that would be able
to know what plugin references to include as deps, but I don't think we need that
quite yet (or maybe ever, we'll see).

Reviewed By: snarkmaster

Differential Revision: D19162032

fbshipit-source-id: 1dada1df846f164938f5c24d960074420a1cd630