Break out repo-server to be a stand alone binary that is shipped with the snapshot.
Summary:
This breaks out `repo-server` to not be embedded as part of `yum-from-snapshot` so that it
can be used in `PEX` style builds.  This is not a permanent solution as the real ideal state would
be for the `repo-server` to live outside the BA, that work is being done, but I need this to unblock OSS progress.

Reviewed By: snarkmaster

Differential Revision: D19185677

fbshipit-source-id: f8d914a1d82e00028adc288bee3371f264339168