Mark `rpm_repo_snapshot` uncacheable
Summary:
I messed up when I removed the `install_buck_runnable` usage from this file.

The failure mode is that we'd fetch `rpm_repo_snapshot` from a remote build cache, but the wrapper inside that cached file would refer to a `yum-from-snapshot` binary that wasn't available in the local `buck-out`.

Reviewed By: zeroxoneb

Differential Revision: D19220818

fbshipit-source-id: 63ecec6d01ae083b1af890f41b24dac60372b8a8