Genericize `yum_conf.py` to support `dnf`
Summary:
This is **mostly** a search-and-replace / rename from "yum" to "yum+dnf".

As far as we are concerned, there are almost no differences between `yum.conf` and `dnf.conf` formats. Some difference in the default directories (duh), and the presence of an overridable `varsdir` in `dnf` is all that I spotted in  a cursory pass.

I "bifurcated" the test so that it runs both branches of the code depending on the subclass (`YumConfTestCase` vs `DnfConfTestCase`).

Reviewed By: zeroxoneb

Differential Revision: D19227075

fbshipit-source-id: 61bf6a362a606886a786d16eb877620760d7bbe4