Add support for snapshotting `dnf.conf`
Summary:
The idea is that a snapshot can contain a `yum.conf`, a `dnf.conf`, or both. In the latter case, they must refer to the same repos (see the rationale in a comment below).

For the purposes of our tests, I make `dnf.conf` almost exactly the same as `yum.conf`, aside from renaming the relevant paths to their `dnf` defaults.

Such a snapshot setup allows us to use the test build appliance either with `dnf` or with `yum`.

Reviewed By: zeroxoneb

Differential Revision: D19227082

fbshipit-source-id: 769889dadbd2052201f3b31332ae8ce09bc65f30