Improve RPM action -> yum/dnf command mapping
Summary:
This is needed to add `dnf` support. There's a subtle difference from `yum`, which is that `install-n /path/to.rpm` no longer works in `dnf`.

I refactored the logic to make it (hopefully) much clearer what's going on.

In doing so I noticed that we weren't handling the case of `image.rpms_remove_if_exists([":target"])`. Up to now, this would tend to result in a silent misbehavior (i.e. `yum` would refuse to remove anything but exit with error code 0).

In this diff I make that simply remove any RPM with the same name as `:target`, which seems actually useful. I added a test to this effect.

Reviewed By: zeroxoneb

Differential Revision: D19207508

fbshipit-source-id: 3416181b9d310a17ad8dc04e4063dae3b9b97fca