`yum-dnf-from-snapshot` actually supports `dnf`; not yet used
Summary:
This adds a `yum_dnf` positional arg to `yum-from-dnf-snapshot`, and fishes it through from `yum.sh` down into `yum-dnf-from-snapshot`.

This is the last of the changes to `yum-dnf-from-snapshot` to make it work with `dnf`. Note that there's *nothing* unexpected here. There's a minor change in the prior D19227115, and there are some minor surprises in the tests (next diff).

The included changes to the  `-from-snapshot` tests:
 - Are intended as no-op refactors.
 - Let the next diff add "dnf" variants via 1-line subclasses, the way we do in `test-yum-dnf-conf` (D19227075).
 - Deliberately omit DNF-specific tweaks to the tests.

To see the difference that switching to `dnf` makes to the tests, go here:
 - D19229267 for the main test
 - [not for commit] D19230581 for the FB build appliance test

Reviewed By: zeroxoneb

Differential Revision: D19229267

fbshipit-source-id: 6c92851e26661c080e8d31b38b4579577789b61d