Run `test-yum-dnf-from-snapshot` in host BA
Summary:
The `dnf`-flavored part of `test-yum-dnf-from-snapshot` (introduced in D19230576) relies on the ambient environment having `/var/lib/dnf`. Running the test inside a BA guarantees this.

Unfortunately, the host BA cannot be guaranteed to contain the versionlock plugins for `dnf` and `yum`, so the `versionlock.gz` hacks will stay for now.

Reviewed By: zeroxoneb

Differential Revision: D19235718

fbshipit-source-id: 1b70fcc5a0a2bc4e8211d033d2e7d29e708d4585