Use UnionFind in repo_sizer
Summary:
In the comments on [this diff](https://our.intern.facebook.com/intern/diff/D18715939/?transaction_id=458870144773102), Alexey mentioned how repo sizer was essentially implementing union find, and we could eventually move the implementation to that.

While addressing his other comments, I discovered another small bug in the implementation, and in the process of fixing it (which was causing the code to become even more bloated), decided it would be worth quickly attempting to implement the union-find method.

It turned out to be relatively simple and made the initial code considerably easier to follow while also causing the merge of two sizers to be trivial, so I think this small change is well worth the benefits.

Reviewed By: snarkmaster

Differential Revision: D19171964

fbshipit-source-id: 10a61f1ad6353aeeb487d46566b54ec039c79ff0