Copy union-find to rpm
Summary:
Copies union-find implementation over to `fs_image` so we can use it in open-source.

Even though not technically third-party I placed it in that directory given that it essentially is for our case here, but let me know if you think it might live in a better place.

Also wasn't entirely sure what we meant when referring to backlinking - I just put a comment referencing where the file came from.

Reviewed By: zeroxoneb

Differential Revision: D19195063

fbshipit-source-id: 0aaea000abcc161421a392dca4898758ca35d14b