Fix test-yum-dnf-from-snapshot test
Summary: This test now runs as an `image_python_unittest`, which means it runs inside an `image.layer`, specifically the fake host build appliance one. What this means is that there is no need to encoded the location of the snapshot or the `repo-server` binary into the test binary itself.  Since it's running in an image it can just use the standard paths for those artifacts.

Reviewed By: rdna

Differential Revision: D19238035

fbshipit-source-id: 5fe23243d4cf73ff1830d35aabe925ad2e1ad18e