Restrict allowed characters in universe name
Summary:
I'm changing all the DB columns that should have human-readable data to be `TEXT` / `VARCHAR` — including `universe`.

The fact that allowable characters for other RPM fields are over-broad and loosely defined makes for some headaches, so let's start with a very restrictive charset for `universe`. This is fine, since we get to make these up.

Reviewed By: zeroxoneb

Differential Revision: D19266754

fbshipit-source-id: 93fbad698e32260164a4f4645067f0a1c014ed0d