Cross-universe mutable RPM detection; reorder DB index
Summary:
This fixes a design glitch with the new `universe` column. A single snapshot potentially collects RPMs from multiple universes. Therefore, it's incorrect to confine mutable RPM detection to just one universe — if we have two possible contents for one NEVRA in a snapshot, even if they came from different universes, it's still a determinism problem, and should be flagged.

Here, I change the logic to look for mutable RPMs in all the universes involved in the snapshot.

This kind of lookup is not very efficient if `universe` is first in the index, since you need to a point lookup for each universe separately. The in-code docs give some other reasons why NEVRA should be first in the index.

For those reason, this diff also changes the primary key order for the `rpm` table.

Reviewed By: zeroxoneb

Differential Revision: D19325960

fbshipit-source-id: 6e67693e1ade635b6d847ce4df7e850566286e35