Make storage/db CLI args deferred for RPM snapshots
Summary:
In the upcoming diffs we attempt to parallelize RPM snapshotting in code, which will be useful for automated snapshots.

Today the arguments provided in the CLI are automatically parsed and a connection object is created. This doesn't work when we're parallelizing as we don't want to create these objects until the threads have been spawned.

Thus, as a solution, I've added some deferral logic so we can defer the construction until a later point, by leveraging `partial`. I thought this was a good middle-ground between added complexity and allowing us to do what we want, without having to modify a ton of the plumbing.

If you want to actually see this in use feel free to take a look at the parallelization diff stacked above.

Update: We've decided to instead just make this the default instead of keeping the ability to have both.

Reviewed By: snarkmaster

Differential Revision: D19295791

fbshipit-source-id: 2e5c0df6585f092816c9d868be0d34c32d31115d