Make sqlite ctx mgr more aggressive
Summary:
Current sqlite context manager keeps the connection open for the lifetime of the object. This was causing problems when parallelizing as we had too many simultaneous open connections and were exhausting the pool. I've instead made it similar to the mysql manager and only open a connection on `__enter__`.

I also had to modify some tests that were using `:memory:` and broke after this change. These broke in two ways:
- Because `:memory:` causes the db to be wiped when connections are closed, now that we're being more aggressive some cases started failing because connections were closed at certain points in the test case - I moved to using a temp file to fix this and have the db persist
 - Tangentially to the above, some test cases did multiple steps in one function. Using `:memory:` inherently caused the db to be wiped after connections were closed, giving a clean slate, which was no longer the case with a file - so I truncate the file between steps

Reviewed By: zeroxoneb

Differential Revision: D19295792

fbshipit-source-id: e4313e59bd3a732f541e46ac772691b0735fadcc