Fix executable wrapper collision bug
Summary:
Before this change, if a single TARGETS file wrapped the same path twice in mode/opt, the second one would go down the `if exists:` branch, and it would incorrectly return a reference to `dummyfile`, a 0-byte non-executable file, instead of just returning the unmodified target path.

The way I found it was that it broke when I put two `install_rpm_repo_snapshot` calls in one project, since both tried to reference `//fs_image/rpm:yum-dnf-from-snapshot`. And in one snapshot, the copied file was empty and non-executable, but only in mode/opt.

Oops.

Reviewed By: justintrudell

Differential Revision: D19463233

fbshipit-source-id: 35dc6281122b9a11f2d355b40e0989fc7029a763