Auto-update image fbpkg versions for oncall 'twimage'
Summary:
**Do not change the title of this diff.**

This automatic update of fbpkg data was triggered by the [Chronos job fs_image.fbcode.update.fbpkg.db](https://our.internmc.facebook.com/intern/chronos/job/?jobname=fs_image.fbcode.update.fbpkg.db&smc=chronos_gp_admin_client), documented on the [Automatic Repo Updates wiki](https://our.intern.facebook.com/intern/wiki/FS_Image/Automatic_Repo_Updates).

This diff will be automatically accepted by the submitting service user and, if all tests are green, will be landed through 'Accept-to-Ship'.

 ---

## I am having trouble landing this diff

If tests are broken, the responding human should address the failure with the owner of the broken contbuild, and ensure that the fix lands the on trunk.  The next run of the Chronos job will update this diff with a fix.

If landing this diff fails due to a merge conflict, it is normally OK to wait until Chronos updates this diff with the next version.

 ---

## My package update has to break compatibility

If a specific package update **must** break compatibility with a specific project, the preferred solution is to create a commit that fixes the project, and then to run:

```
fbpkg tag YOUR_PACKAGE:NEW_UUID YOUR_TAG
buck run //fs_image/fbpkg/facebook/db:update-db \
  --db fs_image/fbpkg/facebook/db/main_db/ \
  --no-update-existing \
  --replace 'YOUR_PACKAGE' 'YOUR_TAG' '{}'
hg amend
```

This way, extraneous changes to the package DB are not coupled with the project update.  The main downside is the risk of a merge conflict on landing the auto-generated diff, but we can fix this it it occurs often.

 ---

## I am mitigating a SEV

In emergencies, it is acceptable to commandeer this diff, and to land the DB update together with fixes to dependent code.  Please note:
 - Prefer to follow the "breaking compatibility" steps above, if possible.
 - Commandeering this diff will cause Chronos to (eventually) create a new one.

 ---

## How was this generated?

The Sandcastle job `fs_image-fbcode-update-fbpkg-db` executed this command:

```
buck run fs_image/fbpkg/facebook/db:publish-main-db-update
```

Use lolbunny `sandcastle fs_image-fbcode-update-fbpkg-db` to inspect recent runs.

Differential Revision: D19551935

fbshipit-source-id: 8479578f5fa1d3bc246f4e66b9852939083aa0ed