Have repo-server use relative storage paths under snapshot_dir when using filesystem
Summary:
We often use a relative storage path in repo-server, but having that relative path work under the current directory is usually not the desired result, so let's always have it be relative to the snapshot_dir.

This snippet was already present elsewhere, so let's just move it to repo-server to ensure it's always the case.

Reviewed By: snarkmaster

Differential Revision: D19564011

fbshipit-source-id: d9dcdfe6031f79d184b791254b22aec136fb4052