Allow nspawn_in_subvol to spawn a repo-server
Summary:
Extend nspawn_in_subvol so it's able to inject a socket into the container's network namespace and run a repo-server attached to that socket.

This will allow us to simplify the working of the build appliance, which right now runs a repo-server (not on a network namespace, since it needs to communicate with the outside) and also runs yum_from_snapshot (in a private network, since it only needs to talk to the repo-server proxy.) With this change, we will be able to run the repo-server outside of the build appliance and consequently have the build_appliance fully inside a private network. Furthermore, this change will allow us to decouple the release of the build appliance with the release of the RPM database snapshot.

This is also a base for running Chef inside Buck image builds, since Chef wants to be able to run yum (or dnf), so it needs to be able to access a yum server, which we can provide (in a hermetic way) with the repo-server proxy.

There are currently a few limitations:

 - Only works for `yum` (but adding support for `dnf` shouldn't be too hard.)
 - Only works for non-booted container (it's the case we need right now.) I have some refactoring planned for the booted container case (using sd-notify to wait for it and get the PID), so I'm investigating whether to build on top of that for the booted case.
 - Only takes a storage configuration and `repo-server` binary from the snapshot directory. This is a restriction in some ways, but on the other hand it makes it easy to use, with a single path to configure everything.

Reviewed By: snarkmaster

Differential Revision: D19050986

fbshipit-source-id: d09a4c8a58d4f01469533ab72dba266ac181acec