Pass config rather than context to repo_downloader
Summary:
Updates plumbing to pass/handle a config dict rather than an already constructed db/storage object, which is needed to parallelize (as we want to make these connections in the individual threads).

Note that reconstructing RepoDBContext this many times is actually a minor efficiency hit because in the `__init__` we check to ensure tables exist. A diff further up in the stack optimizes this (I didn't do it here as many of these contexts get removed anyways when parallelizing).

Reviewed By: snarkmaster

Differential Revision: D19523627

fbshipit-source-id: cefa7eecf7793bab1ab105db5af8399031996b7c