Break repo downloader out of class
Summary:
Bulk changes here are taking the methods from `repo_downloader` out of a class and instead passing parameters as arguments. This is a necessary step to enabling the parallelization of the downloads. The large majority of changes in that file are essentially just removing an indent for everything and dropping `self`.

The `download()` method has also become `download_repos()` and operates on a sequence of repos. To be able to accomplish this, we also need to include `repo` in `RepoSnapshot` (to know which repo a snapshot corresponds to), and this was also included.

There are minimal functional changes in this diff (I've tried to explicitly highlight them), it's moreso in preparation for the functional changes in the upcoming diffs.

Reviewed By: snarkmaster

Differential Revision: D19668781

fbshipit-source-id: 19c7de61142bfbd01d96dc2004165c02ab446c5a