Reimplement retries in downloader
Summary:
This reimplements the `retry_fn` functionality that used to be in `snapshot_repos` to instead happen in the downloader.

It will become clearer in the threading diffs why this makes sense as a decorator, but I also think it offers a clean overall implementation.

It is absolutely possible to keep `retries` as a command-line argument and extract it from the args in the decorator, but I feel we don't really gain much by having it on the command-line (it's kind of inherent to the execution model here and I can't see a strong case why you'd want to disable it - it's rarely kicked off anyways). If you disagree I'm happy to move this around, but for now I've just made it a static arg.

With all of the upcoming changes to the download model overall retries in general may become obsolete, but we can revisit that once we get all of this moved forward.

Reviewed By: zeroxoneb, snarkmaster

Differential Revision: D19668817

fbshipit-source-id: 550b61818b8a328b6f8d29eff5e2c33acd4c92c2