Refactor third_party.library shim
Summary: This refactors the `third_party.library(..)` macro for the oss_shim to use kwargs instead of
trying to do string parsing.  This has the added effect of meaning that we can kill the
`external_deps` usage in at least the python targets.  The c++ targets still need external_deps,
but in OSS those don't work yet anyway, so more work will be done in follow ups to make this work
for c++.

Reviewed By: snarkmaster

Differential Revision: D19996476

fbshipit-source-id: 75ce19de1c4d2c3a22adceecf282707efe5b92bc