OSS implementation of third_party targets (#7)
Summary:
Initial implementation of supporting `third_party` libraries and targets for OSS builds.

This add both the `oss_shim.bzl` implementation for OSS to resolve a target for a `third-party` library and the targets themselves.  This is starting point for using these kinds of libs in OSS, there is certainly more work to do, but this gets builds for `rpm` and other subsystems working.
Pull Request resolved: https://github.com/facebookincubator/fs_image/pull/7

Reviewed By: snarkmaster

Differential Revision: D19968569

Pulled By: zeroxoneb

fbshipit-source-id: e3dbc5ed2938b608bb730d047b3f6191fc3edefe