Add `fetched_with_nondeterministic_fs_metadata` package targets
Summary:
(Also see the in-code docs)

This adds a "for advanced users only" variant of `_fetched_package_layer`. The variant is called `_fetched_package_with_nondeterministic_fs_metadata`.

It gives the user the option **not** to create an `image.layer` for the content of the package. This means that the contents goes in `buck-out`, and can only represent:
 - regular files and directories
 - "user can execute bits"

No other fancy VFS features (e.g. "owners" or "xattrs") are expected to work in this setup.

The new gadget's name is purposefully long, a bit obscure, and a bit threatening. We want normal people to use fetched layers and know exactly what they're getting.

There are a couple of situations in which it's advantageous to use this instead of `_fetched_package_layer`:
 - In a world where ~all operations go through the build appliance, we need to be able to fetch the build appliance's sendstream without instantiating an `image.layer` for it. That breaks a circular dependency that mpatlasov ran into.
 - Your package contains something insensitive to VFS niceties (such as a compressed sendstream) **and** you don't want to re-fetch the package every time something about the `fs_image` compiler changes. This makes it a nice "developer efficiency" boost for images that are used as part of the core `fs_image` developer experience.

As a drive-by, I dropped support for `.bzl` package databases, since we weren't using them, and weren't likely to be using them.

Reviewed By: zeroxoneb

Differential Revision: D20105534

fbshipit-source-id: a99c2c47ccff29562f4ef5e8f6e1a6e84742afad