Support configurable targets for validating host mount features
Summary:
This makes it so that the allowed targets that can be used to define host mounts is configurable. It's explicitly *not* doing a startswith match so that each and every host_mount target *must* be defined.  Since host mounts are considered extermely dangerous, being more restrictive and explicit about which *exact* targets are allowed to use them seems like the safest approach. This will have the affect of causing an explicit failure if a host_mount is unintentionally added without the config being update.  The hope is that this kind of change would be caught in testing/CI and trigger a human interaction.

The configuration mechanism is to use `.buckconfig` and set the list of targets in the `fs_image.allowed_host_mount_targets` config.

Note: there's a new test to validate that deeply nested anonymous targets can't be used for host_mounts.  This should probably be fixed so that anonymous features get their source target properly attributed, but that should happen in another change.

Reviewed By: snarkmaster

Differential Revision: D20099900

fbshipit-source-id: 6ffd3d32661405fe9c74641e52311eb62a52c6b8