Split repo_downloader into sub-modules
Summary:
Splits the large `repo_downloader.py` into separate modules as it's become fairly bloated, and each operation is relatively independent from the other. (definitely should've done this earlier in the stack but ya win some ya lose some).

Made a sub-dir `downloader` to house all of these, and also moved `deleted_mutable_rpms` into this as it's essentially tied to the downloader.

As part of this, also moved the `retryable` decorator to the `rpm/common` as was recommended in a previous diff.

Reviewed By: zeroxoneb

Differential Revision: D19977419

fbshipit-source-id: cbbb5024b2533f7fdf94e9c8394a208cf31cccfe