Parallelize repodata downloads
Summary:
Out of all diffs in the stack, this is likely the one with the most functional changes.

On top of parallelizing the downloads of repodatas, we also change the recovery behaviour considerably, such that we now store repodatas in the DB as we download them, regardless of if the overall operation succeeds or not. The justification is mentioned in a comment but the gist is that the cleanup logic adds a lot of extra complexity, and as long as the top-level object referencing them (repomd) isn't committed, they'll just be extra objects in the db with a negligible overall storage footprint. If we run into storage issues with this we can also easily add a periodic clean-up job to delete any unreferenced objects.

You'll also notice I've opted to perform the writes serially from the master thread. This is because sqlite doesn't handle writing concurrency well, and I was still running into issues where snapshots would sometimes fail to get the lock and crash. Rather than spend a lot of time debugging this, it made more sense to just write serially, especially considering this amounts to a small percentage of our overall execution time (downloads dominate that).

Reviewed By: zeroxoneb

Differential Revision: D19726825

fbshipit-source-id: 710709def299ac409fd95bd33d2ff35988bf1b78