Read-only db connections in downloader
Summary:
This is a general security benefit to ensure we're not accidentally writing in the downloader.

However, my hopes are that this will also have an effect on the performance of the job in the SC. Some digging around, and recent results (e.g. see [the logs](https://our.intern.facebook.com/intern/sandcastle/log/?instance_id=27021597981982370&step_id=27021599789453571&step_index=4&name=Snapshot+RPMs+and+BA%2C+create+diff) for this recently failed job) have led me to believe that at least some of the slowness of our job in SC is reading from the repo db for already stored IDs when processing RPMs.

Naturally we always connect to the db in write mode currently, which means we are never using the replicas and only connecting to the master. We'd also unnecessarily be waiting on locks whenever one thread is writing a new storage ID.

I suspect there's still a lot more going on under the hood causing slowness but my hopes are this will potentially help (and as I said it's a security win regardless).

Reviewed By: snarkmaster

Differential Revision: D20176512

fbshipit-source-id: 1e9a45ae06cf783e0eb689d55aa66fd850a6e78c