Coerce tuples to lists in nspawn-in-subvol's repo-server support.
Summary:
The current code fails if tuples are passed, since `[:]` doesn't turn them into lists and the default for forward_fd in new_nspawn_ops is now an empty tuple.

Use the `list` constructor instead, which will copy an existing list if one is passed, but also coerce a tuple to a list if required.

Reviewed By: snarkmaster

Differential Revision: D20197544

fbshipit-source-id: b59af7b38a1ee3e075cd73328053bbcd08581da4