Retry failed db connections
Summary:
New issue with automated snapshots is that our db connections are failing because they're idle for more than 10 seconds ([see logs](https://our.intern.facebook.com/intern/sandcastle/log/?instance_id=22517998356515827&step_id=22518000172517180&step_index=4&name=Snapshot+RPMs+and+BA%2C+create+diff)).

I audited our code and it looks like we properly open and close a connection as tightly as possible, so I don't think stale connections are the issue. I'd imagine it's potentially just due to client starvation if we have too much contention going on.

Figured an easy way to try to get around this is just to implement retries on the place that seems to be most error-prone, which is the heavy read path in RPMs.

Reviewed By: snarkmaster

Differential Revision: D20224943

fbshipit-source-id: bf52e9fd166c3b58fed27cc0d80294d276b76982