Use dataclasses and regular class inheritance for ImageItem
Summary:
We should get rid of use of metaclasses and extended namedtuples in favor of standard Python libraries that are more friendly to type checking.

While type checking is a driver for this change, it was long overdue to drop some of the unnecessary complexity of our codebase.

In order to handle the massaging of our inputs, we are using a custom constructor rather than use the one from `dataclass(init=False)`, but in this code we're reusing the internals of dataclasses used to expose this constructor. We're handling the API changes from upstream, so we'll be ready for the known upcoming change to it.

I ran the typechecker on this code, posting a follow up for it.

Reviewed By: snarkmaster

Differential Revision: D20291545

fbshipit-source-id: 70986f9a3a8185942121021186c4778de866254b