Populate the license in each source file as required by OSS policy
Summary:
This is a huge diff that touches just about every file.  The only thing here is the insertion of the license header and a few modifications in places where we reference BSD instead of MIT (origianlly I think we planned for BSD but went with MIT instead).

Changes applied with the linter using the changes in the diff below this D20350894

Reviewed By: snarkmaster

Differential Revision: D20330307

fbshipit-source-id: b346d003c2e4d0da4845182dd79ed3a59d843f20