Clean up write paths in repo snapshots
Summary:
While investigating the snapshot failure caused by DB consistency I noticed our writes paths can be improved a bit in snapshotting.

Now that we're entirely in-process and perform our writes from the master thread, we don't need to worry about racing writers causing a different storage ID to be persisted, and can remove the "maybe" semantics of `maybe_write_id`. I've also updated this to have the writing function no longer accept ReportableErrors and expect the client to deal with that, as I figured that went along with the idea of removing "maybe" semantics.

This won't change behaviour but IMO helps with readability.

Reviewed By: zeroxoneb

Differential Revision: D20346193

fbshipit-source-id: a2983af71ba9e544775f7383ad7d7a687992d97e