Fedora 31 Kernel RPMs (#12)
Summary:
### Goal
The goal of this change is to be closer to creating a default "kernel artifact" (a kernel bundle containing `vmlinuz`, `initrd`, `modules`, `headers`, `devel`) which is used to run a "vmtest". In fact, the only component that is missing from the list above is the `initrd` because it is more tightly coupled with fb-internal logic.

### Changes
Similar to how the `busybox` rpm is made available, I have added a few buck targets for `kernel-{core,header,devel}` rpms which are part of the Fedora 31 repository. These are then used to create/extract `modules` and `vmlinuz`.
Pull Request resolved: https://github.com/facebookincubator/fs_image/pull/12

Test Plan:
```
➜  kernel git:(fedora31-kernel-rpms) ✗ buck clean && buck fetch //third-party/fedora31/kernel/...Starting new Buck daemon...
Parsing buck files: finished in 4.2 sec
Building: finished in 21.4 sec (100%) 3/3 jobs, 3 updated
  Total time: 26.1 sec

➜  kernel git:(fedora31-kernel-rpms) ✗ buck build //third-party/fedora31/kernel/...
69800 blocks
Parsing buck files: finished in 1.9 sec
Building: finished in 17.3 sec (100%) 6/6 jobs, 3 updated
  Total time: 19.3 sec
```

### Next Step
The next step will be an fb-internal change which will move much of what is in `kernel/vm`, including the `initrd` build macro ,into fs_image. In doing so, we can then use to macro in fs_image to completely constuct our "kernel artifact" from scratch in OSS.

Reviewed By: zeroxoneb

Differential Revision: D20368298

Pulled By: NathanJiangCS

fbshipit-source-id: c06f4c96a89e0d2e70c06687064291dfa07d2642