Use dataclasses instead of enriched namedtuple in btrfs_diff
Summary:
Similarly here, using a superclass to define the common attributes (`path`) and methods.

The type annotations are likely incorrect but we can address those later, as we enable type checking on this part of the tree.

This particular case exposed a dependency we have on the `__repr__` format, which we use for some error messages and we have tests checking those in a regex. The fact that we have these in inner classes changes the `__repr__` result in dataclasses, so we're overriding it here to stay compatible with the original format we expected.

Reviewed By: snarkmaster

Differential Revision: D20314493

fbshipit-source-id: 78f4bf0e8f7dd8462f00589cfd18342ea42daf5a