Dispatch `image_*_unittest` command-line rewriting by test type
Summary: This is a no-op refactor to make it more explicit what we are doing, with the intent of making it easy to support more test runners in the future. As a side effect, we will no longer be running command-rewriting logic for gtests, where it didn't make sense anyway.

Differential Revision: D20395163

fbshipit-source-id: 8b3b6ff9ae46ce43ddf7d378f1fb7424c98c8c3f